Create EncCodeGroup
I got this from here http://www.fourmilab.ch/javascrypt/jscrypt.html they said it was in the public domain but I couldn't find it anywhere so i took it from the website.

What I want to do is to shorten the length of the final encrypted message. I also want to include numbers instead of just letters in the final encrypted message. The hope is to take about 250 characters of text in the un-encrypted box and make the final encrypted message no more that 27 groups of 5 characters each if possible. I'd like to shorten the final encrypted message as small as I can possibly get it.